Extract mainEventId to cjs .js file, reference in next.config.js
Allows de-duplicating mainEvent ID.
