+++ web/util/constants.js
@@ -1,0 +1,4 @@
+// This file needs to be CJS due to it being referenced by next.config.js.
+module.exports = {
+  mainEventId: 'aad77280-6394-4090-afad-1c0f2a0416c6',
+};
